Hatch 348/v2 datetime frontend (#231)
* HATCH-348: frontend support for datetime v2

* merge in feat branch

* merge feat

* HATCH-348: v2 datetime + v2 update/create/hooks

* deleteOne & useDeleteOne

* component typing

* fix types & tests

+++ packages/rest-client/src/services/promise/deleteOne/deleteOne.ts
@@ -1,15 +1,28 @@
-import type { Schemas, Source } from "../../types"
+import type {
+  FinalSchemas,
+  GetSchemaNames,
+  PartialSchemas,
+  Source,
+} from "../../types"
 import { notifySubscribers } from "../../store"
+import { SchemaNameNotStringError, schemaNameIsString } from "../../utils"
 
 /**
  * Deletes a resource in the data source and notifies subscribers.
  */
-export const deleteOne = async (
+export const deleteOne = async <
+  const TSchemas extends PartialSchemas,
+  const TSchemaName extends GetSchemaNames<TSchemas>,
+>(
   dataSource: Source,
-  allSchemas: Schemas,
-  schemaName: string,
+  allSchemas: FinalSchemas,
+  schemaName: TSchemaName,
   id: string,
 ): Promise<void> => {
+  if (!schemaNameIsString(schemaName)) {
+    throw new SchemaNameNotStringError(schemaName)
+  }
+
   await dataSource.deleteOne(allSchemas, schemaName, id)
 
   notifySubscribers()
